# src/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Principal Loja7",
    "version": "1.0.0",
    "description": "API de acesso principal."
  },
  "paths": {
    "/produtos": {
      "get": {
        "summary": "Acessa a API de Produtos.",
        "description": "Encaminha as requisições GET para a API de Produtos. <br><a href=\"/produtos\" target=\"_blank\">Acessar Documentação da API de Produtos</a>",
        "responses": {
          "302": {
            "description": "Redirecionamento para a interface Swagger da API de Produtos"
          },
          "404": {
            "description": "API de Produtos não encontrada ou indisponível."
          },
          "500": {
            "description": "Erro ao comunicar com a API de Produtos."
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Acessa a API de Autenticação de Usuários.",
        "description": "Encaminha as requisições GET para a API de Autenticação de Usuários. <br><a href=\"/usuarios\" target=\"_blank\">Acessar Documentação da API de Usuários</a>",
        "responses": {
          "302": {
            "description": "Redirecionamento para a interface Swagger da API de Autenticação de Usuários"
          },
          "404": {
            "description": "API de Autenticação de Usuários não encontrada ou indisponível."
          },
          "500": {
            "description": "Erro ao comunicar com a API de Autenticação de Usuários."
          }
        }
      }
    },
    "/pedidos": {
      "get": {
        "summary": "Acessa a API de Pedidos.",
        "description": "Encaminha as requisições GET para a API de Pedidos. <br><a href=\"/pedidos\" target=\"_blank\">Acessar Documentação da API de Pedidos</a>",
        "responses": {
          "302": {
            "description": "Redirecionamento para a interface Swagger da API de Pedidos"
          },
          "404": {
            "description": "API de Pedidos não encontrada ou indisponível."
          },
          "500": {
            "description": "Erro ao comunicar com a API de Pedidos."
          }
        }
      }
    },
    "/requests_count": {
      "get": {
        "summary": "Retorna um objeto JSON contendo o número de requisições GET para os endpoints /usuarios, /produtos e /pedidos.",
        "description": "Obtém a contagem de requisições para os endpoints principais. <br><a href=\"/requests_count\" target=\"_blank\">Acessar contagem de visitantes</a>",
        "responses": {
          "200": {
            "description": "Contagem de requisições obtida com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "pedidos": {
                  "type": "integer",
                  "example": 4
                },
                "produtos": {
                  "type": "integer",
                  "example": 2
                },
                "usuarios": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          },
          "404": {
            "description": "Endpoint para contagem de requisições não encontrado."
          },
          "500": {
            "description": "Erro ao obter a contagem de requisições."
          }
        }
      }
    }
  }
}